Split Classes. Fix clone quick
- Separated a bunch of UI classes and Task classes into a classes subfolder for better management
- Added a DRY_RUN option to test without needing to pull
- Added a hasattr check for "git_dir" in Repository as sometimes it was causing issues
- Fixed an issue where some repo(s) could be incorrectly handled and never finalised, likely due to extremely fast processing. Added a sleep in task runner in task_queue.py after starting the task to attempt to mitigate this.
