+++ src/conf_globals.py
@@ -5,3 +5,4 @@ APP_NAME: str = "PyGitDatBack"
 COMMIT_CUTOFF_DAYS = 360
 THREAD_TIMEOUT_SECONDS = 60
 MAX_CONCURRENT_TASKS = 3
+DRY_RUN = False
