+++ src/ui/__init__.py
@@ -1,3 +1,1 @@
-__VERSION__ = (0, 0, 1)
-
 from .ui_main import GitDatBackUI
