+++ src/ui/classes/worker_signals.py
@@ -1,0 +1,6 @@
+from PySide6.QtCore import QObject, Signal
+
+class WorkerSignals(QObject):
+    finished = Signal(str)
+    error = Signal(str, str)
+
